requirements.txt
transformers
datasets
torch
fastapi
uvicorn
sentry-sdk
python-dotenv
